backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.30
django-filter==24.3
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.5
typing-extensions==4.13.2
